Park Miller is working again
I'm pretty sure it has to have a 64-bit state, just going by what I've seen with the basic LCG. I think the examples I saw online are wrong. The schrage method shouldn't require a 64-bit variable, but the regular Park Miller should. When it was set to 32-bit state, the first couple numbers in the sequence were right of course, but as soon as it overflowed it yielded different numbers. Might try to compile a C++ program or something to compare against...

+++ Pseudorandom/PRNG.h
@@ -111,10 +111,10 @@ uint32_t PRNG_NextBounded(PRNG *self, uint32_t lower, uint32_t upper);
 // -----------------------------------------------------------------------------
 
 uint32_t LCGBig_Next(uint64_t *state);
-uint16_t LCGSmall_Next(uint32_t *state); // not implemented yet
+uint16_t LCGSmall_Next(uint32_t *state);
 
 // uint32_t ParkMillerBigger_Next(uint64_t *state); // not implemented yet
-uint32_t ParkMiller_Next(uint32_t *state);
+uint32_t ParkMiller_Next(uint64_t *state);
 
 uint32_t Schrage_Next(uint32_t *state);
